arch: move file-backed stores and bus server to infra layer (#290)
Move three modules from app/ to infra/ where they belong:
- app/task.rs → infra/task_store.rs (file-backed task queue)
- app/statemachine.rs → infra/sm_store.rs (file-backed SM store)
- app/bus.rs → infra/bus_server.rs (Unix socket bus server)

Original app modules now re-export from infra for backward compat —
all existing `crate::app::task::*` paths continue to work.
