+++ src/infra/mod.rs
@@ -1,8 +1,11 @@
 //! Infrastructure layer — concrete implementations of port traits.
 
+pub mod bus_server;
 pub mod context_store;
 pub mod dto;
 pub mod memory_bus;
 pub mod memory_store;
 pub mod paths;
+pub mod sm_store;
+pub mod task_store;
 pub mod unix_bus;
